fix: change index and size data types from int to size_t
This way, huge images of 2000x2000x2000 dimensions won't cause overflow.

+++ include/morphology/compare_arrays_grayscale.h
@@ -5,14 +5,14 @@
 
 template <typename dtype>
 void compare_arrays_grayscale(dtype* deviceImage1, dtype* deviceImage2, int* deviceOutput,
-                              const int size, const int flag_verbose);
+                              const size_t size, const int flag_verbose);
 
 template <typename dtype>
 void compare_arrays_grayscale_on_device(dtype* hostImage1, dtype* hostImage2, int* hostOutput,
-                                        const int size, const int flag_verbose);
+                                        const size_t size, const int flag_verbose);
 
 template <typename dtype>
 void compare_arrays_grayscale_on_host(dtype* hostImage1, dtype* hostImage2, int* hostOutput,
-                                      const int size);
+                                      const size_t size);
 
 #endif  // COMPARE_ARRAYS_GRAYSCALE_H
