+++ src/morphology/grayscale/morph_chain_grayscale.cu
@@ -31,7 +31,7 @@ void morph_chain_grayscale_on_device(dtype* hostImage, dtype* hostOutput, const 
                                      int kernel_xsize, int kernel_ysize, int kernel_zsize,
                                      MorphChain chain) {
   // set input dimension
-  int size = xsize * ysize * zsize;
+  size_t size = xsize * ysize * zsize;
   size_t nBytes = size * sizeof(dtype);
   size_t nBytes_padding = xsize * ysize * (padding_bottom + padding_top) * sizeof(dtype);
 
@@ -124,7 +124,7 @@ void morph_chain_grayscale_on_host(dtype* hostImage, dtype* hostOutput, const in
                                    int kernel_ysize, int kernel_zsize, MorphChain chain) {
 
   // set input dimension
-  int size = xsize * ysize * zsize;
+  size_t size = xsize * ysize * zsize;
   size_t nBytes = size * sizeof(dtype);
 
   // allocate temporary memory
